+++ packages/utils/src/hooks/index.ts
@@ -1,0 +1,5 @@
+import mountMPA from './mountMPA/mountMPA';
+import mountSPA from './mountSPA/mountSPA';
+import parseRequest from './parseRequest/parseRequest';
+
+export { mountMPA, mountSPA, parseRequest };