+++ packages/utils/src/index.ts
@@ -1,0 +1,5 @@
+import { mountMPA, mountSPA } from './hooks/index';
+
+export { mountMPA, mountSPA };
+
+export type * from './hooks/types';